Update CoreNLP module to use relative paths
The Stanford CoreNLP module of choice requires a path to the Stanford
CoreNLP jars for their models and general execution logic. For now
we've decided to use relative paths and to check the jars in as
libraries to the project. This could be abstracted out to environment
variables in the future.